requirements/compiled.txt
MySQL-python==1.2.3c1
Jinja2==2.5.5

# for bcrypt passwords
hmac==20101005
hashlib==20081119
py-bcrypt==0.2

python-ldap==2.3.13
